[refactor][css] v2.0.0 frontend 구조화 변경 ① Header 영역 hooks 분리 REF : #11
